refactor(duelling-picklist): convert stories to typescript
